:sparkles: feat: Implementa método para alterar usuário

+++ src/main/java/com/traveler/api/repository/ViagemRepository.java
@@ -3,5 +3,5 @@ package com.traveler.api.repository;
 import com.traveler.api.entity.Viagem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ViagemRepository extends JpaRepository<Viagem, Integer> {
+public interface ViagemRepository extends JpaRepository<Viagem, Long> {
 }
